tutorials/vllm-samples/deepseek/eks/fraud-detection-demo/requirements.txt
# MCP Server Dependencies
mcp>=0.9.0
fastmcp>=0.2.0
uvicorn>=0.30.0
pydantic>=2.0.0

# AWS SDK
boto3>=1.34.0

# HTTP requests
requests>=2.31.0
aiohttp>=3.9.0

# Strands Agent Framework
strands-agents>=1.0.0
strands-agents-tools>=0.2.0
